feat(schematic): refactor main heirarchy
This commit makes a bunch of changes to how power is delivered on the
schematic. It also adds the probit header and the i2c output connector
on the main page of the schematic. Lastly, it changes it so that USB
power should work and also not power the 5V rail to the mosfet ICs.